usando modulo de network e convertendo JSON

+++ features/search/data/src/main/java/io/github/josebatista/marketplace/data/model/Attribute.kt
@@ -13,6 +13,6 @@ internal data class Attribute(
     @SerialName("attribute_group_name") val attributeGroupName: String? = null,
     @SerialName("value_struct") val valueStruct: ValueStruct? = null,
     @SerialName("values") val values: List<AttributeValue>? = null,
-    @SerialName("source") val source: Int? = null,
+    @SerialName("source") val source: Long? = null,
     @SerialName("value_type") val valueType: String? = null,
 )
