+++ features/search/data/src/main/java/io/github/josebatista/marketplace/data/model/AttributeValue.kt
@@ -8,5 +8,5 @@ internal data class AttributeValue(
     @SerialName("id") val id: String? = null,
     @SerialName("name") val name: String? = null,
     @SerialName("struct") val struct: ValueStruct? = null,
-    @SerialName("source") val source: Int? = null,
+    @SerialName("source") val source: Long? = null,
 )
